+++ Tests/JSONTests/JSONTests.swift
@@ -13,105 +13,6 @@ func perf(_ block: () throws -> Void) rethrows -> Double {
     return time
 }
 
-extension StructuredData {
-    /**
-     Attempt to initialize a node with a foundation object.
-
-     - warning: will default to null if unexpected value
-     - parameter any: the object to create a node from
-     - throws: if fails to create node.
-     */
-    public init(any: Any) {
-        switch any {
-            // If we're coming from foundation, it will be an `NSNumber`.
-        //This represents double, integer, and boolean.
-        case let number as Double:
-            // When coming from ObjC Any, this will represent all Integer types and boolean
-            self = .number(Number(number))
-        // Here to catch 'Any' type, but MUST come AFTER 'Double' check for JSON fuzziness
-        case let bool as Bool:
-            self = .bool(bool)
-        case let int as Int:
-            self = .number(Number(int))
-        case let uint as UInt:
-            self = .number(Number(uint))
-        case let string as String:
-            self = .string(string)
-        case let object as [String : Any]:
-            self = StructuredData(any: object)
-        case let array as [Any]:
-            self = .array(array.map(StructuredData.init))
-        case _ as NSNull:
-            self = .null
-        case let data as Data:
-            self = .bytes(data.makeBytes())
-        case let bytes as NSData:
-            var raw = [UInt8](repeating: 0, count: bytes.length)
-            bytes.getBytes(&raw, length: bytes.length)
-            self = .bytes(raw)
-        case let date as Date:
-            self = .date(date)
-        case let date as NSDate:
-            let date = Date(timeIntervalSince1970: date.timeIntervalSince1970)
-            self = .date(date)
-        default:
-            self = .null
-        }
-    }
-
-    /**
-     Initialize a node with a foundation dictionary
-     - parameter any: the dictionary to initialize with
-     */
-    public init(any: [String: Any]) {
-        var mutable: [String: StructuredData] = [:]
-        any.forEach { key, val in
-            mutable[key] = StructuredData(any: val)
-        }
-        self = .object(mutable)
-    }
-
-    /**
-     Initialize a node with a foundation array
-     - parameter any: the array to initialize with
-     */
-    public init(any: [Any]) {
-        let array = any.map(StructuredData.init)
-        self = .array(array)
-    }
-
-    /**
-     Create an any representation of the node,
-     intended for Foundation environments.
-     */
-    public var any: Any {
-        switch self {
-        case .object(let ob):
-            var mapped: [String : Any] = [:]
-            ob.forEach { key, val in
-                mapped[key] = val.any
-            }
-            return mapped
-        case .array(let array):
-            return array.map { $0.any }
-        case .bool(let bool):
-            return bool
-        case .number(let number):
-            return number.double
-        case .string(let string):
-            return string
-        case .null:
-            return NSNull()
-        case .bytes(let bytes):
-            var bytes = bytes
-            let data = NSData(bytes: &bytes, length: bytes.count)
-            return data
-        case .date(let date):
-            return date
-        }
-    }
-}
-
 class CompareTests: XCTestCase {
     static let allTests = [
         ("testJayPerformance", testJayPerformance),
@@ -150,10 +51,6 @@ class JSONTests: XCTestCase {
     static let allTests = [
         ("testParse", testParse),
         ("testSerialize", testSerialize),
-        ("testComments", testComments),
-        ("testCommentsSingle", testCommentsSingle),
-        ("testCommentsInternal", testCommentsInternal),
-        ("testCrazyCommentInternal", testCrazyCommentInternal),
         ("testSerializePerformance", testSerializePerformance),
         ("testParsePerformance", testParsePerformance),
     ]
@@ -194,53 +91,14 @@ class JSONTests: XCTestCase {
         XCTAssert(serialized.contains("\"array\":[null,true,1337,\"😄\"]"))
     }
 
-    func testComments() throws {
-        let string = " /* asdfg */ {\"1\":1}"
-        do {
-            let parsed = try JSON(serialized: string, allowComments: true)
-            XCTAssertEqual(parsed["1"]?.int, 1)
-        } catch {
-            XCTFail("Could not parse: \(error)")
-        }
-    }
-
-    func testCommentsSingle() throws {
-        let string = " {\"1\":1 // test \n }"
-        do {
-            let parsed = try JSON(serialized: string, allowComments: true)
-            XCTAssertEqual(parsed["1"]?.int, 1)
-        } catch {
-            XCTFail("Could not parse: \(error)")
-        }
-    }
-
-    func testCommentsInternal() throws {
-        let string = " {\"1\":\"/* comment */\"}"
-        do {
-            let parsed = try JSON(serialized: string, allowComments: true)
-            XCTAssertEqual(parsed["1"]?.string, "/* comment */")
-        } catch {
-            XCTFail("Could not parse: \(error)")
-        }
-    }
-
-    func testCrazyCommentInternal() throws {
-        let string = "{\"1\": \"Here's a great comment quote \\\"/*why are people doing this*/\\\"\"}"
-        do {
-            let parsed = try JSON(serialized: string, allowComments: true)
-            XCTAssertEqual(parsed["1"]?.string, "Here's a great comment quote \"/*why are people doing this*/\"")
-        } catch {
-            XCTFail("Could not parse: \(error)")
-        }
-    }
-
     func testPrettySerialize() throws {
         let json = try JSON(node: [
             "hello": "world"
         ])
 
         let serialized = try json.serialize(prettyPrint: true).string
-        XCTAssertEqual(serialized, "{\n    \"hello\": \"world\"\n}")
+        let expectation = "{\n  \"hello\" : \"world\"\n}"
+        XCTAssertEqual(serialized, expectation)
     }
 
     func testStringEscaping() throws {
